Added all basic methods to controllers and services

+++ src/main/java/nl/moreniekmeijer/backendspringboottechiteasycontroller/repositories/WallBracketRepository.java
@@ -3,6 +3,5 @@ package nl.moreniekmeijer.backendspringboottechiteasycontroller.repositories;
 import nl.moreniekmeijer.backendspringboottechiteasycontroller.models.WallBracket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WallBracketRepository extends JpaRepository<WallBracket, Integer> {
-
+public interface WallBracketRepository extends JpaRepository<WallBracket, Long> {
 }
